+++ include/caster/caster_cpu.h
@@ -7,19 +7,19 @@ using namespace std;
 
 class CasterCPU : public Caster {
  public:
-  CasterCPU(int n, function<void(float)> onErr, function<void(vector<float2>&)> onPos)
+  CasterCPU(int n, function<void(double)> onErr, function<void(vector<double2>&)> onPos)
       : Caster(n, onErr, onPos) {}
   ~CasterCPU(){};
 
   virtual void simul_step() override {
     simul_step_cpu();
     if (it++ % 100 == 0) {
-      float err = 0.0;
+      double err = 0.0;
       for (auto& dist : distances) {
-        float d = dist.r;
-        float2 iPos = positions[dist.i];
-        float2 jPos = positions[dist.j];
-        float2 ij = {iPos.x - jPos.x, jPos.y - jPos.y};
+        double d = dist.r;
+        double2 iPos = positions[dist.i];
+        double2 jPos = positions[dist.j];
+        double2 ij = {iPos.x - jPos.x, jPos.y - jPos.y};
         err += abs(d - sqrt(ij.x * ij.x + ij.y * ij.y));
       }
       onError(err);
