+++ include/caster/caster_cuda.h
@@ -6,23 +6,23 @@
 using namespace std;
 
 struct Sample {
-  float2 pos;
-  float2 v;
-  float2 *components;
+  double2 pos;
+  double2 v;
+  double2 *components;
   short num_components;
 };
 
 class CasterCuda : public Caster {
  public:
-  CasterCuda(int n, function<void(float)> onErr, function<void(vector<float2>&)> onPos)
+  CasterCuda(int n, function<void(double)> onErr, function<void(vector<double2>&)> onPos)
       : Caster(n, onErr, onPos) {}
 
   void sortHostSamples(vector<int> &labels);
 
-  float2 *d_positions;
+  double2 *d_positions;
   DistElem *d_distances;
   Sample *d_samples;
-  float *d_errors;
+  double *d_errors;
 
   virtual void prepare(vector<int> &labels) override;
   virtual void finish() override;
@@ -34,7 +34,7 @@ class CasterCuda : public Caster {
  protected:
   void initializeHelperVectors();
   virtual void simul_step_cuda() = 0;
-  virtual float getError();
+  virtual double getError();
   void copyPositions();
   int itToPosReady = -1;
   unsigned it = 0;
