+++ include/caster/caster_cuda_adam.h
@@ -4,8 +4,8 @@ using namespace std;
 
 class CasterCudaAdam : public CasterCuda {
  public:
-  CasterCudaAdam(int n, function<void(float)> onErr,
-                 function<void(vector<float2>&)> onPos)
+  CasterCudaAdam(int n, function<void(double)> onErr,
+                 function<void(vector<double2>&)> onPos)
       : CasterCuda(n, onErr, onPos) {}
   virtual void prepare(vector<int> &labels) override;
 
@@ -14,5 +14,5 @@ class CasterCudaAdam : public CasterCuda {
 
  private:
   // x,y -> squared mean, z,w -> mean
-  float4 *d_average_params;
+  double4 *d_average_params;
 };
