+++ src/caster/caster_cuda_adam.cu
@@ -8,13 +8,13 @@ using namespace std;
 #define EPS 0.00000001f
 #define LEARNING_RATE 0.002f
 
-__global__ void calcPositionsAdam(long n, unsigned it, Sample *samples, float4 *avarage_params) {
+__global__ void calcPositionsAdam(long n, unsigned it, Sample *samples, double4 *avarage_params) {
   for (int i = blockIdx.x * blockDim.x + threadIdx.x; i < n;
       i += blockDim.x * gridDim.x) {
     Sample sample = samples[i];
-    float4 avarage_param = avarage_params[i];
+    double4 avarage_param = avarage_params[i];
 
-    float2 force = {0, 0};
+    double2 force = {0, 0};
     for (int j = 0; j < sample.num_components; j++) {
       force.x += sample.components[j].x;
       force.y += sample.components[j].y;
@@ -26,8 +26,8 @@ __global__ void calcPositionsAdam(long n, unsigned it, Sample *samples, float4 *
     avarage_param.z = avarage_param.z * B1 + (1.0 - powf(B1,it)) * force.x;
     avarage_param.w = avarage_param.w * B1 + (1.0 - powf(B1,it)) * force.y;
 
-    float deltax = LEARNING_RATE * (avarage_param.z / (1.0 - B1)) / (EPS + sqrtf(avarage_param.x / (1.0 - B2)));
-    float deltay = LEARNING_RATE * (avarage_param.w / (1.0 - B1)) / (EPS + sqrtf(avarage_param.y / (1.0 - B2)));
+    double deltax = LEARNING_RATE * (avarage_param.z / (1.0 - B1)) / (EPS + sqrtf(avarage_param.x / (1.0 - B2)));
+    double deltay = LEARNING_RATE * (avarage_param.w / (1.0 - B1)) / (EPS + sqrtf(avarage_param.y / (1.0 - B2)));
 
     sample.pos.x += deltax;
     sample.pos.y += deltay;
@@ -44,18 +44,18 @@ __global__ void calcForceComponentsAdam(int compNumber, DistElem *distances,
       i += blockDim.x * gridDim.x) {
     DistElem distance = distances[i];
 
-    float2 posI = samples[distance.i].pos;
-    float2 posJ = samples[distance.j].pos;
+    double2 posI = samples[distance.i].pos;
+    double2 posJ = samples[distance.j].pos;
 
-    float2 rv = posI;
+    double2 rv = posI;
     rv.x -= posJ.x;
     rv.y -= posJ.y;
 
-    float r = sqrtf((posI.x - posJ.x) * (posI.x - posJ.x) +
+    double r = sqrtf((posI.x - posJ.x) * (posI.x - posJ.x) +
         (posI.y - posJ.y) * (posI.y - posJ.y) + 0.00001f);
-    float D = distance.r;
+    double D = distance.r;
 
-    float energy = (r - D) / r;
+    double energy = (r - D) / r;
     rv.x *= -energy;
     rv.y *= -energy;
 
@@ -79,6 +79,6 @@ void CasterCudaAdam::prepare(vector<int> &labels) {
   CasterCuda::prepare(labels);
 
   // TODO release this memory
-  cuCall(cudaMalloc(&d_average_params, positions.size() * sizeof(float4)));
-  cuCall(cudaMemset(d_average_params, 0, positions.size() * sizeof(float4)));
+  cuCall(cudaMalloc(&d_average_params, positions.size() * sizeof(double4)));
+  cuCall(cudaMemset(d_average_params, 0, positions.size() * sizeof(double4)));
 }
